# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DisciplineX API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Task {
  id: string;
  user_id: string;
  title: string;
  /** @nullable */
  description?: string | null;
  completed: boolean;
  /** @nullable */
  completed_at?: string | null;
  /** @nullable */
  due_date?: string | null;
  /** @nullable */
  priority?: string | null;
  created_at: string;
  updated_at?: string;
}

export type TaskInputPriority = typeof TaskInputPriority[keyof typeof TaskInputPriority];


export const TaskInputPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface TaskInput {
  user_id: string;
  /** @minLength 1 */
  title: string;
  description?: string;
  due_date?: string;
  priority?: TaskInputPriority;
}

export interface TaskUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  due_date?: string;
  priority?: string;
}

export interface TaskCompleteInput {
  completed: boolean;
}

export interface UserStats {
  user_id: string;
  discipline_score: number;
  weekly_change?: number;
  current_streak: number;
  longest_streak: number;
  total_tasks_completed: number;
  weekly_completion_pct: number;
  level: string;
  momentum_score: number;
  /** @nullable */
  last_active_date?: string | null;
  updated_at?: string;
}

export interface UserStatsInput {
  discipline_score?: number;
  weekly_change?: number;
  current_streak?: number;
  longest_streak?: number;
  total_tasks_completed?: number;
  weekly_completion_pct?: number;
  level?: string;
  momentum_score?: number;
  last_active_date?: string;
}

export interface DailyActivity {
  id: string;
  user_id: string;
  date: string;
  tasks_completed: number;
  tasks_total: number;
  discipline_delta?: number;
}

export interface ActivityInput {
  user_id: string;
  tasks_completed: number;
  tasks_total: number;
  discipline_delta?: number;
}

export interface DisciplineResult {
  discipline_score: number;
  level: string;
  weekly_change: number;
  momentum_score: number;
  current_streak?: number;
}

export type ListTasksParams = {
user_id: string;
/**
 * @nullable
 */
date?: string | null;
/**
 * @nullable
 */
completed?: boolean | null;
};

export type FetchActivityParams = {
user_id: string;
/**
 * @nullable
 */
days?: number | null;
};