# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DisciplineX API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: tasks
    description: Task operations
  - name: stats
    description: User stats and discipline engine
  - name: activity
    description: Daily activity tracking
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks for current user
      parameters:
        - in: query
          name: user_id
          required: true
          schema:
            type: string
        - in: query
          name: date
          required: false
          schema:
            type: ["string", "null"]
        - in: query
          name: completed
          required: false
          schema:
            type: ["boolean", "null"]
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  /tasks/{id}/complete:
    patch:
      operationId: completeTask
      tags: [tasks]
      summary: Toggle task completion
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCompleteInput"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

  /stats/{user_id}:
    get:
      operationId: getUserStats
      tags: [stats]
      summary: Get user discipline stats
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserStats"
    put:
      operationId: upsertUserStats
      tags: [stats]
      summary: Upsert user stats
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserStatsInput"
      responses:
        "200":
          description: Updated stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserStats"

  /activity:
    get:
      operationId: fetchActivity
      tags: [activity]
      summary: Get daily activity for heatmap
      parameters:
        - in: query
          name: user_id
          required: true
          schema:
            type: string
        - in: query
          name: days
          required: false
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Activity records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyActivity"
    post:
      operationId: recordActivity
      tags: [activity]
      summary: Record activity for today
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityInput"
      responses:
        "200":
          description: Activity recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyActivity"

  /stats/{user_id}/discipline:
    post:
      operationId: recalculateDiscipline
      tags: [stats]
      summary: Recalculate discipline score
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated discipline score
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DisciplineResult"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Task:
      type: object
      required: [id, user_id, title, completed, created_at]
      properties:
        id:            { type: string }
        user_id:       { type: string }
        title:         { type: string }
        description:   { type: ["string", "null"] }
        completed:     { type: boolean }
        completed_at:  { type: ["string", "null"] }
        due_date:      { type: ["string", "null"] }
        priority:      { type: ["string", "null"] }
        created_at:    { type: string }
        updated_at:    { type: string }

    TaskInput:
      type: object
      required: [user_id, title]
      properties:
        user_id:       { type: string }
        title:         { type: string, minLength: 1 }
        description:   { type: string }
        due_date:      { type: string }
        priority:      { type: string, enum: [low, medium, high] }

    TaskUpdate:
      type: object
      properties:
        title:         { type: string, minLength: 1 }
        description:   { type: string }
        due_date:      { type: string }
        priority:      { type: string }

    TaskCompleteInput:
      type: object
      required: [completed]
      properties:
        completed: { type: boolean }

    UserStats:
      type: object
      required: [user_id, discipline_score, current_streak, longest_streak, total_tasks_completed, weekly_completion_pct, level, momentum_score]
      properties:
        user_id:                { type: string }
        discipline_score:       { type: number }
        weekly_change:          { type: number }
        current_streak:         { type: integer }
        longest_streak:         { type: integer }
        total_tasks_completed:  { type: integer }
        weekly_completion_pct:  { type: number }
        level:                  { type: string }
        momentum_score:         { type: number }
        last_active_date:       { type: ["string", "null"] }
        updated_at:             { type: string }

    UserStatsInput:
      type: object
      properties:
        discipline_score:       { type: number }
        weekly_change:          { type: number }
        current_streak:         { type: integer }
        longest_streak:         { type: integer }
        total_tasks_completed:  { type: integer }
        weekly_completion_pct:  { type: number }
        level:                  { type: string }
        momentum_score:         { type: number }
        last_active_date:       { type: string }

    DailyActivity:
      type: object
      required: [id, user_id, date, tasks_completed, tasks_total]
      properties:
        id:               { type: string }
        user_id:          { type: string }
        date:             { type: string }
        tasks_completed:  { type: integer }
        tasks_total:      { type: integer }
        discipline_delta: { type: number }

    ActivityInput:
      type: object
      required: [user_id, tasks_completed, tasks_total]
      properties:
        user_id:          { type: string }
        tasks_completed:  { type: integer }
        tasks_total:      { type: integer }
        discipline_delta: { type: number }

    DisciplineResult:
      type: object
      required: [discipline_score, level, weekly_change, momentum_score]
      properties:
        discipline_score: { type: number }
        level:            { type: string }
        weekly_change:    { type: number }
        momentum_score:   { type: number }
        current_streak:   { type: integer }