# lib/api-zod/src/generated/types/activityInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DisciplineX API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityInput {
  user_id: string;
  tasks_completed: number;
  tasks_total: number;
  discipline_delta?: number;
}