# lib/api-zod/src/generated/types/disciplineResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DisciplineX API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DisciplineResult {
  discipline_score: number;
  level: string;
  weekly_change: number;
  momentum_score: number;
  current_streak?: number;
}