# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DisciplineX API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Task {
  id: string;
  user_id: string;
  title: string;
  /** @nullable */
  description?: string | null;
  completed: boolean;
  /** @nullable */
  completed_at?: string | null;
  /** @nullable */
  due_date?: string | null;
  /** @nullable */
  priority?: string | null;
  created_at: string;
  updated_at?: string;
}