# lib/api-zod/src/generated/types/taskCompleteInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DisciplineX API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TaskCompleteInput {
  completed: boolean;
}