# lib/api-zod/src/generated/types/taskInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DisciplineX API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TaskInputPriority } from './taskInputPriority';

export interface TaskInput {
  user_id: string;
  /** @minLength 1 */
  title: string;
  description?: string;
  due_date?: string;
  priority?: TaskInputPriority;
}