# lib/api-zod/src/generated/types/userStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DisciplineX API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserStats {
  user_id: string;
  discipline_score: number;
  weekly_change?: number;
  current_streak: number;
  longest_streak: number;
  total_tasks_completed: number;
  weekly_completion_pct: number;
  level: string;
  momentum_score: number;
  /** @nullable */
  last_active_date?: string | null;
  updated_at?: string;
}